% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{as_complex}
\alias{as_complex}
\title{Converting to complex by casting dim into real/imaginary}
\usage{
as_complex(x, .margin = 2)
}
\arguments{
\item{x}{an object of class \code{matrix} or \code{array}.}

\item{.margin}{the dimension to interpret as real and imaginary numbers}
}
\value{
complex vector, matrix, or array
}
\description{
Converting to complex by casting dim into real/imaginary
}
