% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD-class.R
\name{as_FD.data.frame}
\alias{as_FD.data.frame}
\title{Convert data.frame to FDboost data format}
\usage{
\method{as_FD}{data.frame}(x, obj.formula, ...)
}
\arguments{
\item{x}{a data.frame including the response values, functional response arguments,
dimension indicator for multidimensional response functions and an id column
indicating the curve id}

\item{obj.formula}{a formula of the form \code{value^dim ~ arg | id} indicating the respective
columns of \code{x}. Missing \verb{^dim} or \verb{|id} are interpreted as only one dimension/id
occuring in the data.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a data.frame in 'irregular' FDboost format \code{FD_irregular}.
}
\description{
Convert data.frame to FDboost data format
}
\details{
A data.frame will basically directly passed to FDboost without much further
transformation. In principle, only the id-column is replaced by a numeric vector
ranging from 1 to length(unique(id)), and the dim-colum is converted to factor.
}
\seealso{
\linkS4class{FDboost(), FD}
}
