% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD-class.R
\name{as_FD.list}
\alias{as_FD.list}
\title{Interpret response given as list for FDboost}
\usage{
\method{as_FD}{list}(x, formula, ...)
}
\arguments{
\item{x}{either a list with an entry per response unit containing the
response data or a list with one element containing the response data indicated
by the left-hand side of the \code{formula}.}

\item{formula}{a formula indicating the element inside of \code{x} to be converted.
If its left-hand side is NULL, \code{x} will be considered in FDboost format. If
the response name is not contained in the list,
the list will be considered the response object itself.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a data.frame in 'irregular' FDboost format
}
\description{
Interpret response given as list for FDboost
}
\details{
This interpreter applies mfInterpret_for_FDboost to all entries of the response
data list, stacks them to a large data.frame and adds an id variable called
'\emph{id}'.
}
\seealso{
FDboost
}
