% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD-class.R
\name{as_FD}
\alias{as_FD}
\title{Convert other data to the FDboost format}
\usage{
as_FD(x, ...)
}
\arguments{
\item{x}{the data object.}

\item{...}{additional arguments passed to methods.}
}
\value{
an object of class \code{FD} with subclass \code{FD_regular} or \code{FD_irregular}.
#' @seealso \linkS4class{FDboost(), FD}
}
\description{
Convert other data to the FDboost format
}
