% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD-class.R
\name{as_FD.tbl_cube}
\alias{as_FD.tbl_cube}
\title{Interpret response given as tbl_cube for FDboost}
\usage{
\method{as_FD}{tbl_cube}(x, obj.formula, ...)
}
\arguments{
\item{x}{a 'tbl_cube' containing the response values as measures and other variables
as dimensions.}

\item{obj.formula}{a formula of the form \code{value^dim ~ arg | id} reflecting the meaning
of the variables. Missing \verb{^dim} or \verb{|id} are interpreted as only one dimension/id
occuring in the data.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a list with a matrix containing the values and vectors of length nrow(matrix)
for the other variables corresponding to the 'regular' FDboost data format
}
\description{
Interpret response given as tbl_cube for FDboost
}
\seealso{
\linkS4class{FDboost(), tbl_cube, FD}
}
