% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_dat-class.R
\name{as_shape_frame_complex}
\alias{as_shape_frame_complex}
\title{Convert objects to \code{shape_frame_complex}}
\usage{
as_shape_frame_complex(x, ...)
}
\arguments{
\item{x}{an object that should be coerced.}

\item{...}{other arguments passed to the individual methods.}
}
\description{
Internal function converting \code{shape_dat} objects to \code{shape_frame_complex}.
As I didn't need it, yet, there's no constructers, validators, or converters
from other formats or form \code{shape_frame_complex} to something else implemented
so far.
}
\seealso{
\code{shape_frame_complex-class}
}
