% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselearner_minus.R
\name{btrafo}
\alias{btrafo}
\title{Baselearner Basis Transformation Operator}
\usage{
btrafo(bl1, Z)
}
\arguments{
\item{bl1}{a baselearner generator object of class \code{blg}.}

\item{Z}{a basis transformation matrix with its number of rows corresponding
to the number of basis functions of \code{bl1}.}
}
\value{
an object of class \code{blg} with the transformed baselearner.
}
\description{
A function transforming the basis of a baselearner linearly with a
basis transformation matrix \eqn{Z}, i.e., if \eqn{B} is the design matrix
of the baselearner, the design matrix of the transformed learner is
\eqn{\tilde{B} = BZ}.
}
