% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docu.R
\docType{data}
\name{cells}
\alias{cells}
\alias{cellr}
\title{Cell outlines simulated from CPM}
\format{
An object of class \code{list} of length 5.

An object of class \code{list} of length 5.
}
\source{
Thankfully provided by Sophia Schaffer.
}
\usage{
data(cells, package = "manifoldboost")

data(cellr, package = "manifoldboost")
}
\description{
2D cell outlines extracted from cells simulated by
the Cellular Potts Model (CPM) proposed by Thueroff et al. 2019, a stochastic
biophysical model, for calibration to real-life cells (Schaffer 2021).
}
\details{
The data is provided as a list with the following entries:
\itemize{
\item \code{a} substrate adhesion (CPM parameter, numeric)
\item \code{r} signaling radius (CPM parameter, numeric)
\item \code{b} bulk stiffness (CPM parameter, numeric)
\item \code{m} membrane stiffness (CPM parameter, numeric)
\item \code{response} A tibble with the cell outline data:
\itemize{
\item \code{id} cell ID (factor, ordered according to covariate vectors)
\item \code{arg} parameterization of the curves (numeric, between 0 and 1)
\item \code{dim} dimension indicator (two-leveled factor)
\item \code{value} x- and y-values of observed points of the outline (numeric)
}
}
}
\references{
Schaffer, S. A. (2021).
Cytoskeletal  dynamics  in  confined  cell  migration:  experiment  andmodelling.
PhD thesis, LMU Munich.  DOI: 10.5282/edoc.28480.

Thueroff, F., A. Goychuk, M. Reiter, and E. Frey (2019).
Bridging the gap betweensingle-cell migration and collective dynamics.
eLife  8, e46842.
}
\keyword{datasets}
