% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfFamily.R
\name{clone,mfboost_family-method}
\alias{clone,mfboost_family-method}
\title{(Deep) clone an S4 class object containing R6 slots}
\usage{
\S4method{clone}{mfboost_family}(object, deep = TRUE)
}
\arguments{
\item{object}{mfFamily object}

\item{deep}{should clone of R6 slots be deep}
}
\value{
cloned object of class mfFamily
}
\description{
(Deep) clone an S4 class object containing R6 slots
}
