% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfFamily.R
\name{clone}
\alias{clone}
\title{(Deep) clone an S4 class object containing R6 slots}
\usage{
clone(object, deep = TRUE)
}
\arguments{
\item{object}{an S4 class object}

\item{deep}{should clone of R6 slots be deep}
}
\value{
cloned version of the object
}
\description{
(Deep) clone an S4 class object containing R6 slots
}
