% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{complex2realmat}
\alias{complex2realmat}
\title{Convert complex to real liner operator matrix}
\usage{
complex2realmat(A)
}
\arguments{
\item{A}{a \code{complex} (or \code{numeric}) matrix.}
}
\description{
\code{complex2realmat()} converts a complex \eqn{n\times m}{(n x m)}-matrix \eqn{A}
representing a
complex linear operator in \eqn{GL(\mathds{C}^m, \mathds{C}^n)}{GL(C^m, C^n)} to a
real \eqn{2n\times 2m}{(2n x 2m)} matrix \eqn{A'} representing the corresponding
real linear
operator in \eqn{GL(\mathds{R}^{2m}, \mathds{R}^{2n})}{GL(R^2m, R^2n)},
identifying \eqn{z \in \mathds{C}^n}{z in C^n} with \eqn{(Re(z^\top), Im(z^\top))^\top}{(Re(z),Im(z)}.
}
