% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{equal_weights}
\alias{equal_weights}
\title{Simple 1/n_grid weights}
\usage{
equal_weights(arg, range = NULL)
}
\arguments{
\item{arg}{A numeric vector with 'time-points' (argument values) the function is evaluated at.
Can be in non-increasing order.}

\item{range}{the range of the interval integrated over. Must include all arg values.}
}
\description{
Constructs a vector with equal weights of 1/n_grid instead of numerical
integration weights for one single function.
}
