% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_FDboost.R, R/factorize_mfboost.R
\name{factorize}
\alias{factorize}
\alias{factorize.FDboost}
\alias{factorize.mfboost}
\title{Factorize tensor product model}
\usage{
factorize(x, ...)

\method{factorize}{mfboost}(
  x,
  blwise = TRUE,
  newdata = NULL,
  newformula = NULL,
  newobj.formula = NULL,
  ...
)
}
\arguments{
\item{x}{a model object of class mfboost.}

\item{...}{other arguments passed to methods (no arguments so far).}

\item{blwise}{logical, should the factorization be carried out base-learner-wise (TRUE, default)
or for the whole model simultaneously.}
}
\value{
a list of two mboost models of class mboost_fac containing basisfunctions
for response and covariates, respectively, as base-learners.

a list of two mboost models of class mboost_fac containing basisfunctions
for response and covariates, respectively, as base-learners.
}
\description{
Factorize an FDboost tensorporduct model into the response and covariate parts for
effect visualization.

Factorize an mfboost tensorporduct model into the response and covariate parts for
effect visualization.
}
\examples{


}
