% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD-class.R
\name{FD-class}
\alias{FD-class}
\alias{FD_regular-class}
\alias{FD_irregular-class}
\title{An \code{FD} S3 class for FDboost data formats}
\description{
The \code{FD} class is an abstract class for lists
containing a sample of 'functional data' objects together
with other variables carrying additional information.
Objects of these class belong either to the subclass \code{FD_regular} or
\code{FD_irregular} reflecting either of the formats desctibed in \code{?FDboost}.
}
\seealso{
\code{\link[=FDboost]{FDboost()}}
}
