% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_FDboost.R
\name{FDboost_fac-class}
\alias{FDboost_fac-class}
\title{\code{FDboost_fac} S3 class for factorized FDboost model components}
\description{
Model factorization with \code{factorize()} decomposes an
\code{FDboost} model into two objects of class \code{FDboost_fac} - one for the
response and one for the covariate predictor. The first is essentially
an \code{FDboost} object and the second an \code{mboost} object, however,
in a 'read-only' mode and slightly adjusted methods (method defaults).
}
\seealso{
\code{\link[=factorize(), factorize.FDboost]{factorize(), factorize.FDboost()}}
}
