% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_dat-class.R
\name{formula.shape_dat}
\alias{formula.shape_dat}
\title{Extract \code{shape_dat} formula}
\usage{
\method{formula}{shape_dat}(x)
}
\arguments{
\item{x}{an object of class \code{shape_dat}.}
}
\value{
a formula
}
\description{
Extracts the formula stating the structure of the shape object in the data.
}
