% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_dat-class.R
\name{get_dat}
\alias{get_dat}
\title{Extract data with structure formula}
\usage{
get_dat(data, formula)
}
\arguments{
\item{data}{an optional \code{data.frame} or \code{list} containing the object structure variables.}

\item{formula}{a formula stating the object structure.}
}
\value{
\code{data.frame} or \code{list} with a \code{formula} attribute.
}
\description{
Creates a object data data.frame from a formula stating the structure of the
object according to a formula template. Used in the \verb{shape_*} functions.
}
