% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselearner_minus.R
\name{\%-\%}
\alias{\%-\%}
\title{Base Learner Basis Orthogonalization Operator}
\usage{
bl1 \%-\% bl2
}
\arguments{
\item{bl1}{a baselearner generator of class \code{blg}}

\item{bl2}{another baselearner generator of class \code{blg} which is typically
contained in \code{bl1} in the sense that the columns of its design matrix are
in the span of the design matrix of \code{bl1}.}
}
\value{
object of class \code{blg} with the transformed baselearner \code{bl1}.
}
\description{
Operator orthogonalizing the design matrix of one \code{mboost} baselearner \code{bl1}
with respect to the design matrix of another baselearner \code{bl2}, i.e.,
the design matrix of \code{bl1} is linearly transformed, such that all columns in
it are orthogonal to the columns of the design matrix of \code{bl2}.
}
