% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.construct.sps.smooth.spec.R
\name{make_summation_matrix}
\alias{make_summation_matrix}
\title{Construct basis transformation matrix for (skew-)symmetry constraints}
\usage{
make_summation_matrix(F, skew = FALSE, cyclic.degree = 0)
}
\arguments{
\item{F}{number of marginal basis functions.}

\item{skew}{logical, should the basis be constraint to skew-symmetry instead
of symmetry.}

\item{cyclic.degree}{integer, specifying the number of basis functions identified
with each other at the boundaries in order to implement periodicity. Should
be specified to match the degree of the B-spline basis.}
}
\value{
A basis transformation matrix of dimension \eqn{F^2 \times G} with
\eqn{G<F^2} depending on the specified constraint.
}
\description{
Construct basis transformation matrix for (skew-)symmetry constraints for
bivariate P-spline smooths, which can also be combined with periodicity constraints
for cyclic marginals. The function is adapted from the
\code{\link[sparseFLMM]{sparseFLMM}} with the same name.
}
\details{
This function is used to implement the respective constraints in the
function \code{\link[smooth.construct.sps.smooth.spec]{smooth.construct.sps.smooth.spec}}.
}
