% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_dat-class.R
\name{match_template}
\alias{match_template}
\title{Match variables from formula to template}
\usage{
match_template(formula, template)
}
\arguments{
\item{formula, template}{formulas of the same form, i.e., equal up to their
variable names.}
}
\value{
A named character vector with the variable names in formula as values
and the corresponding variable names in the template as names.
}
\description{
Internal functions matching the variables in \code{formula} to those provided in a \code{template}.
If formulas do not have the same form, the functions returns an error.
}
