% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_mfboost.R
\name{mfboost_fac-class}
\alias{mfboost_fac-class}
\title{\code{mfboost_fac} S3 class for factorized mfboost model component}
\description{
Model factorization with \code{factorize()} decomposes an
\code{mfboost} model into an object of class \code{mfboost_fac} for the
response and and object of class \code{FDboost_fac} for the covariate predictor.
The first is essentially an \code{mfboost} object and the second an \code{mboost} object,
however,
in a 'read-only' mode and slightly adjusted methods (method defaults).
}
\seealso{
\code{\link[=factorize(), factorize.mfboost(), factorize.FDboost]{factorize(), factorize.mfboost(), factorize.FDboost()}}
}
