% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfFamily.R
\docType{class}
\name{mfboost_family-class}
\alias{mfboost_family-class}
\title{S4 class for a gradient boosting family with geometry slot.}
\description{
The class \code{mfboost_family} inherits from \code{\link[boost_family]{boost_family}},
the \code{mboost} family class. In particular, it's used to define the negative
gradient fitted throughout the boosting algorithm, which might now also depend
on the underlying geometry of the response.
}
\section{Slots}{

\describe{
\item{\code{mf}}{an \code{\link[mfGeometry]{mfGeometry}} object specifying the
underlying geometry.}

\item{\code{pole}}{an \code{\link[mfPole]{mfPole}} object containing a pole model.}

\item{\code{ngradient,loss,risk,response,nuisance}}{functions defining the
\code{mboost::boost_family} object.}

\item{\code{weights,name}}{\code{boost_family} slots indicating the weights allowed and
the name of the loss function.}

\item{\code{offset,fW,rclass,check_y,type}}{other slots of \code{boost_family} objects
that typically do not play a role in \code{mfboost_family} objects.}
}}

