% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfFamily.R
\name{mfFamily}
\alias{mfFamily}
\alias{RiemannL2}
\alias{PlanarShapeL2}
\alias{PlanarSizeShapeL2}
\alias{EuclideanL2}
\title{Manifold Gradient Boosting Families}
\usage{
mfFamily(
  mf,
  pole,
  ngradient,
  update_formula = function(timeformula, pole_) timeformula,
  loss = NULL,
  risk = NULL,
  response = function(f) NA,
  nuisance = function() return(NA),
  weights = c("any", "none", "zeroone", "case"),
  name = "user-specified mfFamily",
  offset = function(y, w) 0,
  fW = NULL
)

RiemannL2(
  mf,
  pole.type = c("RiemannL2", "Gaussian"),
  pole.control = boost_control()
)

PlanarShapeL2(
  pole.type = "RiemannL2",
  pole.control = boost_control(),
  weight_fun = NULL,
  arg_range = NULL
)

PlanarSizeShapeL2(
  pole.type = "RiemannL2",
  pole.control = boost_control(),
  weight_fun = NULL,
  arg_range = NULL
)

EuclideanL2(
  cyclic = FALSE,
  arg.grid.len = 5000,
  weights = weights,
  arg.range = NULL,
  smoothed.cov = NULL,
  cov.k = 10
)
}
\arguments{
\item{mf}{an \code{\link[mfGeometry]{mfGeometry}} object specifying the
underlying geometry.}

\item{pole}{a function with arguments \code{formula} and \code{data}
(in a format used by FDboost) returning the pole in the format required by \code{mf}.}

\item{ngradient, loss, risk, response, nuisance}{functions defining the
\code{mboost::boost_family} object.}

\item{update_formula}{a function with arguments \code{formula} and \code{pole_}
updating the formula in dependence of the \code{pole_}, typically by applying a
function to the right side.}

\item{weights, name}{\code{boost_family} character strings indicating the weights
allowed and the name of the loss function.}

\item{offset, fW, rclass, check_y, type}{other slots of \code{boost_family} objects
that typically do not play a role in \code{mfboost_family} objects.}

\item{pole.type}{one of "RiemannL2"(default) and "Gaussian"}

\item{initialize}{a function with arguments \code{formula} (in obj.formula
format) and \code{data} (in a format used by FDboost) initializing \code{mf}
by setting \code{mf$y_} and \code{mf$weights_}.}

\item{...}{further arguments passed to the}
}
\description{
\code{mfboost_family} objects extend \code{boost_family} objects to contain,
i.a., a geometry slot, which is used to transfer geometric structure information
of the response into the functions defined in a \code{boost_family}.
It is used to specify loss, risk and other functions for fitting regression
problems with manifold valued response with \code{mfboost}
in a flexible and modular way.
}
