% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfInterpreter.R
\name{mfFormulae_for_FDboost.formula}
\alias{mfFormulae_for_FDboost.formula}
\title{Interpret obj.formula as FDboost timeformula}
\usage{
\method{mfFormulae_for_FDboost}{formula}(x, formula, ...)
}
\arguments{
\item{x}{an obj.formula of the form \code{value^dim ~ arg | id} reflecting the structure
of the response. \verb{^dim} and/or \verb{|id} can be missing, when no dimensional/id
information has to be passed.}

\item{formula}{a formula describing the model on a meta level with for the response
object}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Interpret obj.formula as FDboost timeformula
}
