% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfInterpreter.R
\name{mfFormulae_for_FDboost}
\alias{mfFormulae_for_FDboost}
\title{Convert obj.formula and friends to FDboost formulae}
\usage{
mfFormulae_for_FDboost(x, ...)
}
\arguments{
\item{x}{an obj.formula needed for interpretation.}

\item{...}{other arguments needed to interpret \code{x}.}
}
\value{
A list containing the \code{formula}, \code{timeformula} and \code{id}-formula passed
to FDboost
}
\description{
Convert obj.formula and friends to FDboost formulae
}
