% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfGeomElastic.R
\name{mfGeomElasticPlanarShape_closed}
\alias{mfGeomElasticPlanarShape_closed}
\title{Elastic Plane Shape Space Geometry}
\description{
Geometry of the elastic shape space identifying planar shapes
with a centered and scaled complex vector on the complex sphere / complex
projective space.
}
\section{Super classes}{
\code{\link[manifoldboost:mfGeometry]{manifoldboost::mfGeometry}} -> \code{\link[manifoldboost:mfGeomEuclidean]{manifoldboost::mfGeomEuclidean}} -> \code{\link[manifoldboost:mfGeomUnitSphere]{manifoldboost::mfGeomUnitSphere}} -> \code{\link[manifoldboost:mfGeomPlanarShape]{manifoldboost::mfGeomPlanarShape}} -> \code{mfGeomElasticPlanarShape_closed}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mfGeomElasticPlanarShape_closed$new()}}
\item \href{#method-register}{\code{mfGeomElasticPlanarShape_closed$register()}}
\item \href{#method-align}{\code{mfGeomElasticPlanarShape_closed$align()}}
\item \href{#method-distance}{\code{mfGeomElasticPlanarShape_closed$distance()}}
\item \href{#method-srv_trafo}{\code{mfGeomElasticPlanarShape_closed$srv_trafo()}}
\item \href{#method-plot}{\code{mfGeomElasticPlanarShape_closed$plot()}}
\item \href{#method-clone}{\code{mfGeomElasticPlanarShape_closed$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="innerprod">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-innerprod}{\code{manifoldboost::mfGeomEuclidean$innerprod()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomUnitSphere" data-id="exp">}\href{../../manifoldboost/html/mfGeomUnitSphere.html#method-exp}{\code{manifoldboost::mfGeomUnitSphere$exp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="get_normal">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-get_normal}{\code{manifoldboost::mfGeomPlanarShape$get_normal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="log">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-log}{\code{manifoldboost::mfGeomPlanarShape$log()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="register_v">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-register_v}{\code{manifoldboost::mfGeomPlanarShape$register_v()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="structure">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-structure}{\code{manifoldboost::mfGeomPlanarShape$structure()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="structure_weights">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-structure_weights}{\code{manifoldboost::mfGeomPlanarShape$structure_weights()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="transport">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-transport}{\code{manifoldboost::mfGeomPlanarShape$transport()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="unstructure">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-unstructure}{\code{manifoldboost::mfGeomPlanarShape$unstructure()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="unstructure_weights">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-unstructure_weights}{\code{manifoldboost::mfGeomPlanarShape$unstructure_weights()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="validate">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-validate}{\code{manifoldboost::mfGeomPlanarShape$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize planar shape geometry with warping
for data given in (long) FDboost format.
Same as in mfGeomPlanarShape but storing also function arguments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomElasticPlanarShape_closed$new(
  data,
  formula,
  weight_fun = trapez_weights,
  arg_range = c(0, 1),
  warp_update
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data in the format used in \code{\link{FDboost}}.}

\item{\code{formula}}{formula describing the internal structure of the data,
intended for the \code{obj.formula} of \code{mfboost}
(interpreted by \code{mfInterpret_objformula}).}

\item{\code{weight_fun}}{a function computing inner product weights for the geometry
taking two arguments: \code{arg}, the argument of the functional data
specified in \code{formula}, \code{range} the range of \code{arg}.}

\item{\code{arg_range}}{the \code{range} supplied to the \code{weight_fun}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register"></a>}}
\if{latex}{\out{\hypertarget{method-register}{}}}
\subsection{Method \code{register()}}{
scale \code{y_} to unit length or
an SRV-trnasform \code{q_} to unit norm, respectively.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomElasticPlanarShape_closed$register(y_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-align"></a>}}
\if{latex}{\out{\hypertarget{method-align}{}}}
\subsection{Method \code{align()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomElasticPlanarShape_closed$align(y_, y0_, warp = self$warp, eps = 0.01)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a complex vector on the sphere, orthogonal to the constant.}

\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distance"></a>}}
\if{latex}{\out{\hypertarget{method-distance}{}}}
\subsection{Method \code{distance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomElasticPlanarShape_closed$distance(y0_, y1_, squared = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}

\item{\code{y1_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-srv_trafo"></a>}}
\if{latex}{\out{\hypertarget{method-srv_trafo}{}}}
\subsection{Method \code{srv_trafo()}}{
Perform SRV-transform of a curve and its inverse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomElasticPlanarShape_closed$srv_trafo(y_, inverse = FALSE, center = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a complex vector on the sphere, orthogonal to the constant.}

\item{\code{inverse}}{logical, should the inverse SRV-transform be conducted.}

\item{\code{center}}{logical, if \code{inverse = TRUE}, should the output be centered.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
default plotting function for shapes of plane curves, plotting \code{y_}
in front of \code{y0_} (after alignment).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomElasticPlanarShape_closed$plot(
  y_,
  y0_ = self$pole_,
  level = c("curve", "SRV"),
  closed = TRUE,
  center = TRUE,
  add_original = FALSE,
  ylab = NA,
  xlab = NA,
  col = "black",
  pch = 19,
  asp = 1,
  xlim,
  ylim,
  yaxt = "n",
  xaxt = "n",
  type = "l",
  y0_par = list(col = "darkgrey", type = type),
  seg_par = list(col = "grey"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a complex vector on the sphere, orthogonal to the constant.}

\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}

\item{\code{level}}{one of 'curve' (default) and 'SRV' indicating
whether curves are to be plotted or their SRV-transforms}

\item{\code{closed}}{logical, should the plotted functions be closed?}

\item{\code{center}}{logical, if \code{level='curve'}, should curves be centered?}

\item{\code{add_original}}{logical, should original curve y_ be displayed besides its aligned version.}

\item{\code{ylab, xlab, xlim, ylim, xaxt, yaxt, asp}}{graphical parameters passed to \code{base::plot}
with modified defaults.}

\item{\code{col, pch, type}}{graphical parameters passed to \code{base::plot} referring to \code{y_}.}

\item{\code{y0_par}}{graphical parameters for \code{y0_}.}

\item{\code{seg_par}}{graphical parameters for line segments connecting \code{y_}
and \code{y0_}.}

\item{\code{...}}{other arguments passed to \code{base::plot}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomElasticPlanarShape_closed$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
