% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfGeometry.R
\name{mfGeometry}
\alias{mfGeometry}
\title{R6 Class implementing manifold geometries}
\description{
An abstract \code{mfGeometry} class implementing a set of methods, including
e.g. Exp- and Log-mappings, that are needed to define \code{mfboost} families
in a concise and modular way. Children of this class
specify the geometry of interest.
The arguments of the methods are vectors, matrices
or arrays possibly holding additional attributes. Children of this class
specify the geometry of interest. Besides these methods, objects
of class \code{mfGeometry} offer an active slot for the 'pole' which serves
as prototype for elements of the manifold.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-structure}{\code{mfGeometry$structure()}}
\item \href{#method-unstructure}{\code{mfGeometry$unstructure()}}
\item \href{#method-unstructure_weights}{\code{mfGeometry$unstructure_weights()}}
\item \href{#method-align}{\code{mfGeometry$align()}}
\item \href{#method-register}{\code{mfGeometry$register()}}
\item \href{#method-distance}{\code{mfGeometry$distance()}}
\item \href{#method-exp}{\code{mfGeometry$exp()}}
\item \href{#method-log}{\code{mfGeometry$log()}}
\item \href{#method-transport}{\code{mfGeometry$transport()}}
\item \href{#method-innerprod}{\code{mfGeometry$innerprod()}}
\item \href{#method-get_normal}{\code{mfGeometry$get_normal()}}
\item \href{#method-validate}{\code{mfGeometry$validate()}}
\item \href{#method-clone}{\code{mfGeometry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-structure"></a>}}
\if{latex}{\out{\hypertarget{method-structure}{}}}
\subsection{Method \code{structure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$structure(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{a numeric vector representing an element of the manifold in an
unstructured 'flattened' way}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unstructure"></a>}}
\if{latex}{\out{\hypertarget{method-unstructure}{}}}
\subsection{Method \code{unstructure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$unstructure(y_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{an object living in the manifold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unstructure_weights"></a>}}
\if{latex}{\out{\hypertarget{method-unstructure_weights}{}}}
\subsection{Method \code{unstructure_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$unstructure_weights(weights_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-align"></a>}}
\if{latex}{\out{\hypertarget{method-align}{}}}
\subsection{Method \code{align()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$align(y_, y0_ = private$.pole_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{an object living in the manifold}

\item{\code{y0_}}{an object living in the manifold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register"></a>}}
\if{latex}{\out{\hypertarget{method-register}{}}}
\subsection{Method \code{register()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$register(y_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{an object living in the manifold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distance"></a>}}
\if{latex}{\out{\hypertarget{method-distance}{}}}
\subsection{Method \code{distance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$distance(y0_, y1_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{an object living in the manifold}

\item{\code{y1_}}{an object living in the manifold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exp"></a>}}
\if{latex}{\out{\hypertarget{method-exp}{}}}
\subsection{Method \code{exp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$exp(v_, y0_ = private$.pole_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v_}}{an object repesenting a tangent vector}

\item{\code{y0_}}{an object living in the manifold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$log(y_, y0_ = private$.pole_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{an object living in the manifold}

\item{\code{y0_}}{an object living in the manifold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transport"></a>}}
\if{latex}{\out{\hypertarget{method-transport}{}}}
\subsection{Method \code{transport()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$transport(v0_, y0_, y1_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v0_}}{an object repesenting a tangent vector}

\item{\code{y0_}}{an object living in the manifold}

\item{\code{y1_}}{an object living in the manifold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-innerprod"></a>}}
\if{latex}{\out{\hypertarget{method-innerprod}{}}}
\subsection{Method \code{innerprod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$innerprod(v0_, v1_ = v0_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v0_}}{an object repesenting a tangent vector}

\item{\code{v1_}}{an object repesenting a tangent vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_normal"></a>}}
\if{latex}{\out{\hypertarget{method-get_normal}{}}}
\subsection{Method \code{get_normal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$get_normal(y0_ = private$.pole_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{an object living in the manifold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$validate(y0_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{an object living in the manifold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
