% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfGeomElastic.R
\name{mfGeomSRV_closed}
\alias{mfGeomSRV_closed}
\title{Square-root-velocity (SRV) Geometry}
\description{
L2 geometry on SRV-transforms of curves
}
\section{Super classes}{
\code{\link[manifoldboost:mfGeometry]{manifoldboost::mfGeometry}} -> \code{\link[manifoldboost:mfGeomEuclidean]{manifoldboost::mfGeomEuclidean}} -> \code{mfGeomSRV}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mfGeomSRV_closed$new()}}
\item \href{#method-structure}{\code{mfGeomSRV_closed$structure()}}
\item \href{#method-unstructure}{\code{mfGeomSRV_closed$unstructure()}}
\item \href{#method-unstructure_weights}{\code{mfGeomSRV_closed$unstructure_weights()}}
\item \href{#method-align}{\code{mfGeomSRV_closed$align()}}
\item \href{#method-srv_trafo}{\code{mfGeomSRV_closed$srv_trafo()}}
\item \href{#method-plot}{\code{mfGeomSRV_closed$plot()}}
\item \href{#method-clone}{\code{mfGeomSRV_closed$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="distance">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-distance}{\code{manifoldboost::mfGeomEuclidean$distance()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="exp">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-exp}{\code{manifoldboost::mfGeomEuclidean$exp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="get_normal">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-get_normal}{\code{manifoldboost::mfGeomEuclidean$get_normal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="innerprod">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-innerprod}{\code{manifoldboost::mfGeomEuclidean$innerprod()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="log">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-log}{\code{manifoldboost::mfGeomEuclidean$log()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="register">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-register}{\code{manifoldboost::mfGeomEuclidean$register()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="register_v">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-register_v}{\code{manifoldboost::mfGeomEuclidean$register_v()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="transport">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-transport}{\code{manifoldboost::mfGeomEuclidean$transport()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="validate">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-validate}{\code{manifoldboost::mfGeomEuclidean$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize plane SRV geometry.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomSRV_closed$new(
  data,
  formula,
  weight_fun = trapez_weights,
  arg_range = c(0, 1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data in the format used in \code{\link{FDboost}}.}

\item{\code{formula}}{formula describing the internal structure of the data,
intended for the \code{obj.formula} of \code{mfboost}
(interpreted by \code{mfInterpret_objformula}).
If a \code{formula} is already stored in the \code{mfGeometry} object,
this is taken as default.}

\item{\code{weight_fun}}{a function computing inner product weights for the geometry
taking two arguments: \code{arg}, the argument of the functional data
specified in \code{formula}, \code{range} the range of \code{arg}.}

\item{\code{arg_range}}{the \code{range} supplied to the \code{weight_fun}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-structure"></a>}}
\if{latex}{\out{\hypertarget{method-structure}{}}}
\subsection{Method \code{structure()}}{
convert 2D vectors (given in long format with dimension
indicator) to complex
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomSRV_closed$structure(y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unstructure"></a>}}
\if{latex}{\out{\hypertarget{method-unstructure}{}}}
\subsection{Method \code{unstructure()}}{
convert complex back to long format numeric
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomSRV_closed$unstructure(y_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unstructure_weights"></a>}}
\if{latex}{\out{\hypertarget{method-unstructure_weights}{}}}
\subsection{Method \code{unstructure_weights()}}{
double inner product weights for complex vectors
to appear in both dimensions in numeric long format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomSRV_closed$unstructure_weights(weights_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weights_}}{numeric vector of inner product weights matching the \code{y_}
as a complex vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-align"></a>}}
\if{latex}{\out{\hypertarget{method-align}{}}}
\subsection{Method \code{align()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomSRV_closed$align(y_, y0_ = self$pole_, closed = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-srv_trafo"></a>}}
\if{latex}{\out{\hypertarget{method-srv_trafo}{}}}
\subsection{Method \code{srv_trafo()}}{
Perform SRV-transform of a curve and its inverse.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomSRV_closed$srv_trafo(y_, inverse = FALSE, center = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inverse}}{logical, should the inverse SRV-transform be conducted.}

\item{\code{center}}{logical, if \code{inverse = TRUE}, should the output be centered.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
default plotting function for shapes of plane curves, plotting \code{y_}
in front of \code{y0_} (after alignment).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomSRV_closed$plot(
  y_,
  y0_ = self$pole_,
  level = c("curve", "SRV"),
  closed = TRUE,
  center = TRUE,
  ylab = NA,
  xlab = NA,
  col = "black",
  pch = 19,
  asp = 1,
  xlim,
  ylim,
  yaxt = "n",
  xaxt = "n",
  type = "l",
  y0_par = list(col = "darkgrey", type = type),
  seg_par = list(col = "grey"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{one of 'curve' (default) and 'SRV' indicating
whether curves are to be plotted or their SRV-transforms}

\item{\code{closed}}{logical, should the plotted functions be closed?}

\item{\code{center}}{logical, if \code{level='curve'}, should curves be centered?}

\item{\code{ylab, xlab, xlim, ylim, xaxt, yaxt, asp}}{graphical parameters passed to \code{base::plot}
with modified defaults.}

\item{\code{col, pch, type}}{graphical parameters passed to \code{base::plot} referring to \code{y_}.}

\item{\code{y0_par}}{graphical parameters for \code{y0_}.}

\item{\code{seg_par}}{graphical parameters for line segments connecting \code{y_}
and \code{y0_}.}

\item{\code{...}}{other arguments passed to \code{base::plot}.}

\item{\code{add_original}}{logical, should original curve y_ be displayed besides its aligned version.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomSRV_closed$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
