% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfGeomElasticPlanarShape.R
\name{mfGeomWarpPlanarShape}
\alias{mfGeomWarpPlanarShape}
\title{Plane Shape Space Geometry with warping alignment}
\description{
Geometry of the elastic shape space identifying planar shapes
with a centered and scaled complex vector on the complex sphere / complex
projective space.
}
\section{Super classes}{
\code{\link[manifoldboost:mfGeometry]{manifoldboost::mfGeometry}} -> \code{\link[manifoldboost:mfGeomEuclidean]{manifoldboost::mfGeomEuclidean}} -> \code{\link[manifoldboost:mfGeomUnitSphere]{manifoldboost::mfGeomUnitSphere}} -> \code{\link[manifoldboost:mfGeomPlanarShape]{manifoldboost::mfGeomPlanarShape}} -> \code{mfGeomWarpPlanarShape}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mfGeomWarpPlanarShape$new()}}
\item \href{#method-align}{\code{mfGeomWarpPlanarShape$align()}}
\item \href{#method-distance}{\code{mfGeomWarpPlanarShape$distance()}}
\item \href{#method-clone}{\code{mfGeomWarpPlanarShape$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="innerprod">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-innerprod}{\code{manifoldboost::mfGeomEuclidean$innerprod()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomUnitSphere" data-id="exp">}\href{../../manifoldboost/html/mfGeomUnitSphere.html#method-exp}{\code{manifoldboost::mfGeomUnitSphere$exp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="get_normal">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-get_normal}{\code{manifoldboost::mfGeomPlanarShape$get_normal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="log">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-log}{\code{manifoldboost::mfGeomPlanarShape$log()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="plot">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-plot}{\code{manifoldboost::mfGeomPlanarShape$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="register">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-register}{\code{manifoldboost::mfGeomPlanarShape$register()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="register_v">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-register_v}{\code{manifoldboost::mfGeomPlanarShape$register_v()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="structure">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-structure}{\code{manifoldboost::mfGeomPlanarShape$structure()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="structure_weights">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-structure_weights}{\code{manifoldboost::mfGeomPlanarShape$structure_weights()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="transport">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-transport}{\code{manifoldboost::mfGeomPlanarShape$transport()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="unstructure">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-unstructure}{\code{manifoldboost::mfGeomPlanarShape$unstructure()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="unstructure_weights">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-unstructure_weights}{\code{manifoldboost::mfGeomPlanarShape$unstructure_weights()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="validate">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-validate}{\code{manifoldboost::mfGeomPlanarShape$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize planar shape geometry with warping
for data given in (long) FDboost format.
Same as in mfGeomPlanarShape but storing also function arguments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomWarpPlanarShape$new(
  data,
  formula,
  weight_fun = NULL,
  arg_range = NULL,
  closed,
  warp_update
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data in the format used in \code{\link{FDboost}}.}

\item{\code{formula}}{formula describing the internal structure of the data,
intended for the \code{obj.formula} of \code{mfboost}
(interpreted by \code{mfInterpret_objformula}).}

\item{\code{weight_fun}}{a function computing inner product weights for the geometry
taking two arguments: \code{arg}, the argument of the functional data
specified in \code{formula}, \code{range} the range of \code{arg}.}

\item{\code{arg_range}}{the \code{range} supplied to the \code{weight_fun}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-align"></a>}}
\if{latex}{\out{\hypertarget{method-align}{}}}
\subsection{Method \code{align()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomWarpPlanarShape$align(
  y_,
  y0_,
  closed = private$closed,
  warp = self$warp,
  eps = 0.01
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a complex vector on the sphere, orthogonal to the constant.}

\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distance"></a>}}
\if{latex}{\out{\hypertarget{method-distance}{}}}
\subsection{Method \code{distance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomWarpPlanarShape$distance(y0_, y1_, squared = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}

\item{\code{y1_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomWarpPlanarShape$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
