% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfGeomElasticPlanarShape.R
\name{mfGeomWarpPlanarShapeSimple}
\alias{mfGeomWarpPlanarShapeSimple}
\title{Planar Shape Space Geometry with simple warping alignment}
\description{
Geometry of the 'classic' shape space identifying planar shapes
with a centered and scaled complex vector on the complex sphere / complex
projective space, yet also with warping alignment.
}
\section{Super classes}{
\code{\link[manifoldboost:mfGeometry]{manifoldboost::mfGeometry}} -> \code{\link[manifoldboost:mfGeomEuclidean]{manifoldboost::mfGeomEuclidean}} -> \code{\link[manifoldboost:mfGeomUnitSphere]{manifoldboost::mfGeomUnitSphere}} -> \code{\link[manifoldboost:mfGeomPlanarShape]{manifoldboost::mfGeomPlanarShape}} -> \code{mfGeomWarpPlanarShapeSimple}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{mfGeomWarpPlanarShapeSimple$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="align">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-align}{\code{manifoldboost::mfGeomEuclidean$align()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="distance">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-distance}{\code{manifoldboost::mfGeomEuclidean$distance()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="innerprod">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-innerprod}{\code{manifoldboost::mfGeomEuclidean$innerprod()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomUnitSphere" data-id="exp">}\href{../../manifoldboost/html/mfGeomUnitSphere.html#method-exp}{\code{manifoldboost::mfGeomUnitSphere$exp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="get_normal">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-get_normal}{\code{manifoldboost::mfGeomPlanarShape$get_normal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="initialize">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-initialize}{\code{manifoldboost::mfGeomPlanarShape$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="log">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-log}{\code{manifoldboost::mfGeomPlanarShape$log()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="plot">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-plot}{\code{manifoldboost::mfGeomPlanarShape$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="register">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-register}{\code{manifoldboost::mfGeomPlanarShape$register()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="register_v">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-register_v}{\code{manifoldboost::mfGeomPlanarShape$register_v()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="structure">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-structure}{\code{manifoldboost::mfGeomPlanarShape$structure()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="structure_weights">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-structure_weights}{\code{manifoldboost::mfGeomPlanarShape$structure_weights()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="transport">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-transport}{\code{manifoldboost::mfGeomPlanarShape$transport()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="unstructure">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-unstructure}{\code{manifoldboost::mfGeomPlanarShape$unstructure()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="unstructure_weights">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-unstructure_weights}{\code{manifoldboost::mfGeomPlanarShape$unstructure_weights()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomPlanarShape" data-id="validate">}\href{../../manifoldboost/html/mfGeomPlanarShape.html#method-validate}{\code{manifoldboost::mfGeomPlanarShape$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomWarpPlanarShapeSimple$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
