% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfPoles.R
\name{mfInterpret_cplx_curves}
\alias{mfInterpret_cplx_curves}
\title{Complex planar curve interpreter}
\usage{
mfInterpret_cplx_curves(formula, data, weights = NULL, arg.range = NULL)
}
\arguments{
\item{formula}{a formula in the form \code{shape ~ value^dim ~ arg | id}.}

\item{data}{data containing the shape in a format convertible to \code{shape_dat}.
Shape should be contained as \code{tbl_cube} (\code{shape_cube}) if data is regular.}

\item{weights}{}

\item{mf}{a mfGeometry object}
}
\value{
a list containing
\itemize{
\item mod0_data: a data.frame with the dim, id, and arg values needed
to set up a basis for the response object
\item weights: a vector/matrix (regular case) or a list of vectors/matrices of
(integration) weights
\item response: an array (regular case) or vector (irregular case) containing the
response values
}
}
\description{
Complex planar curve interpreter
}
