% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfInterpreter.R
\name{mfInterpret_objformula}
\alias{mfInterpret_objformula}
\title{Extract structural info from obj.formula}
\usage{
mfInterpret_objformula(x, varnames = TRUE, ...)
}
\arguments{
\item{x}{a formula.}

\item{varnames}{logcial, should character variable names be returned (default).
Alternatively formula parts are directly passed.}
}
\value{
a named list containing the \code{value}, \code{dim}, \code{arg} and \code{id}
extracted from the formula.
}
\description{
Extract info from obj.formula relevant for the basic structure of the (response) object.
}
