% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfFamily.R
\name{mfPole}
\alias{mfPole}
\title{R6 Class for mfboost pole model object}
\description{
A pole (similar to an offset) for a \code{mfboost} model is typically result
of some kind of model fit, which might be another simpler mfboost model
or some other kind of model procedure estimating the pole from the data.
For later model predictions on new data, it is important to store the
not only the resulting pole, but the whole model behind it, to have access
to new pole predictions. The abstract R6 class \code{mfPole} implements
the basic structure needed to provide a pole for mfboost.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fit}{\code{mfPole$fit()}}
\item \href{#method-predict}{\code{mfPole$predict()}}
\item \href{#method-clone}{\code{mfPole$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfPole$fit(obj.formula, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfPole$predict(newdata = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfPole$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
