% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfboost.R
\name{plot.mfboost}
\alias{plot.mfboost}
\title{Plot effect estimates of mfboost model}
\usage{
\method{plot}{mfboost}(x, which = NULL, ids = 1:6, multiplier = 1, ...)
}
\arguments{
\item{x}{object of class \code{mfboost} for plotting}

\item{ids}{response ids to be plotted.}

\item{...}{other arguments passed to plot functions.}
}
\value{

}
\description{
Illustrate estimated effects by comparing predictions with corresponding
reference poles.
}
