% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.construct.sps.smooth.spec.R
\name{Predict.matrix.sympspline.smooth}
\alias{Predict.matrix.sympspline.smooth}
\title{Predict matrix method for (skew-)symmetric smooths}
\usage{
\method{Predict.matrix}{sympspline.smooth}(object, data)
}
\arguments{
\item{object}{a \code{sympspline.smooth} object created by
\code{\link[smooth.construct.sps.smooth.spec]{smooth.construct.sps.smooth.spec}}.
See \code{\link[smooth.construct]{smooth.construct}}.}

\item{data}{a list containing the data.
See \code{\link[smooth.construct]{smooth.construct}}.}
}
\value{
A matrix which will map the parameters associated with the smooth to
the vector of values of the smooth evaluated at the covariate values given in object.
}
\description{
Method generating the predict matrix for a (skew-)symmetric (cyclic) univariate
or bivariate P-spline smooth. Adapted from
\code{\link{sparseFLMM::Predict.matrix.symm.smooth}}.
}
