% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiemannL2simulator.R
\name{RiemannL2sim}
\alias{RiemannL2sim}
\title{R6 class for a simulation of RiemannL2 models}
\description{
This class is created to set up a simulations scenario, simulate data from it,
fit models, do cross-validation, predict quantities of interest, and summarize
relevant statistics in an easy and transparent way.
}
\details{
The methods depend on the magrittr \code{\%>\%} operator and \code{dplyr::bind_rows}.
And, if messaging is desired, on the R-package \code{telegram}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RiemannL2sim$new()}}
\item \href{#method-gridsampler}{\code{RiemannL2sim$gridsampler()}}
\item \href{#method-sample_topdown}{\code{RiemannL2sim$sample_topdown()}}
\item \href{#method-sample_blockwise}{\code{RiemannL2sim$sample_blockwise()}}
\item \href{#method-parallel_setup}{\code{RiemannL2sim$parallel_setup()}}
\item \href{#method-parallel_stop}{\code{RiemannL2sim$parallel_stop()}}
\item \href{#method-fit}{\code{RiemannL2sim$fit()}}
\item \href{#method-crossvalidate}{\code{RiemannL2sim$crossvalidate()}}
\item \href{#method-predict}{\code{RiemannL2sim$predict()}}
\item \href{#method-summarize}{\code{RiemannL2sim$summarize()}}
\item \href{#method-notify}{\code{RiemannL2sim$notify()}}
\item \href{#method-plot}{\code{RiemannL2sim$plot()}}
\item \href{#method-clone}{\code{RiemannL2sim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$new(model0, newdata0 = NULL, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridsampler"></a>}}
\if{latex}{\out{\hypertarget{method-gridsampler}{}}}
\subsection{Method \code{gridsampler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$gridsampler(x, size)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_topdown"></a>}}
\if{latex}{\out{\hypertarget{method-sample_topdown}{}}}
\subsection{Method \code{sample_topdown()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$sample_topdown(
  n_sim = 1,
  n_obj = 1,
  n_grid = 40,
  seed = NULL,
  sample_obj = sample,
  sample_grid = self$gridsampler
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_blockwise"></a>}}
\if{latex}{\out{\hypertarget{method-sample_blockwise}{}}}
\subsection{Method \code{sample_blockwise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$sample_blockwise(
  n_sim = 1,
  n_blocks = 1,
  n_grid = 40,
  seed = NULL,
  random_trafo = function(y_) y_,
  sample_grid = self$gridsampler
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parallel_setup"></a>}}
\if{latex}{\out{\hypertarget{method-parallel_setup}{}}}
\subsection{Method \code{parallel_setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$parallel_setup(nc = detectCores())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parallel_stop"></a>}}
\if{latex}{\out{\hypertarget{method-parallel_stop}{}}}
\subsection{Method \code{parallel_stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$parallel_stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$fit(family = self$model0$family, ..., verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crossvalidate"></a>}}
\if{latex}{\out{\hypertarget{method-crossvalidate}{}}}
\subsection{Method \code{crossvalidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$crossvalidate(
  type = "kfold",
  B = 5,
  ...,
  verbose = FALSE,
  seeds = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$predict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$summarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-notify"></a>}}
\if{latex}{\out{\hypertarget{method-notify}{}}}
\subsection{Method \code{notify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$notify(token = bot_token("RBot"), id = user_id("me"), name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiemannL2sim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
