% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_dat-class.R
\name{shape_names}
\alias{shape_names}
\title{Extract \code{shape_dat} stored variable names}
\usage{
shape_names(x)
}
\arguments{
\item{x}{an object of class \code{shape_dat}.}
}
\value{
a named character vector.
}
\description{
Extracts the variable names of the shape object stored in the data.
}
