% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.construct.sps.smooth.spec.R
\name{smooth.construct.sps.smooth.spec}
\alias{smooth.construct.sps.smooth.spec}
\title{(Skew-)Symmetric smooths constructor}
\usage{
\method{smooth.construct}{sps.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth specfication object, usually generated by a term of
the form \code{s(x, bs = "sps", ...)}.}

\item{data}{a list containing the data.}

\item{knots}{a list containing any knots for basis setup. Can be \code{NULL}.}
}
\value{
An object of class "\code{sympspline.smooth}". See
\code{\link[smooth.construct]{smooth.construct}} for its elements.
}
\description{
The \code{sps} smoother class for \code{mgcv} is adapted from the \code{symm}
class in the R package \code{sparseFLMM}. Besides symmetric bivariate P-spline
tensor product smooths, it also offers the opportunity to specify skew-symmetric
smooths, the symmetry constraints also in combination with cyclic (periodic)
marginal P-spline smooths, and the corresponding constraints for univariate
smooths.
}
