% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docu.R
\docType{data}
\name{cells}
\alias{cells}
\title{Cell outlines simulated from CPM}
\format{
...:
\describe{
\item{\code{response}}{A data.frame ...}
}
}
\source{
Generated and simulated by Sophia Schaffer.
}
\usage{
data(cells, package = "manifoldboost")
}
\description{
Outlines extracted from 2D cells simulated from the stochastic
biophysical model, a Cellular Potts Model (CPM), designed by ...
}
\details{
More details
}
\examples{
data("cells", package = "manifoldboost")
}
\references{
Schaffer, Sophia 2021 ...
}
\seealso{
...
}
\keyword{datasets}
