% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfboost.R
\name{mfboost}
\alias{mfboost}
\title{Model-based Boosting for Manifold Valued Object Data}
\usage{
mfboost(formula, obj.formula = NULL, data = NULL, family = Gaussian(), ...)
}
\arguments{
\item{formula}{a symbolic description of the model formula on covariate
level, i.e. specified as if the response was scalar.}

\item{obj.formula}{intrinsic model formula for the internal representation
of the response. E.g., for a functional response \eqn{y_i(t)} typically use
\code{~ bbs(t)} to obtain smooth effects over \eqn{t}. Typically a one-sided
formula, except if each response observation is provided as a data frame. Then
the left-hand side of the formula indicates the variable the response values
are stored in.}

\item{data}{a data frame or list containing the variables in the model.
The response should be either provided as \code{tbl_cube} (regular case) or
as list of \code{tbl_cube}s or \code{data.frame}s (irregular case). See details.}

\item{family}{an \code{\link[mfFamily]{mfFamily}} object.}

\item{...}{additional arguments passed to \code{\link[FDboost]{FDboost}}/\code{\link[mboost]{mboost}}.}

\item{dim}{typically one-sided formula with the dimension indicator for
multidimensional response (curves). For each dimension the \code{obj.formula}
is repeated (or the \code{formula} if \code{obj.formula} is not specified).}
}
\value{
An object of class \code{mfboost} inheriting from \code{FDboost} and
\code{mboost}.
}
\description{
An interface for model-based gradient boosting when response observations
are naturally represented as vectors, matrices,
or smooth potentially multidimensional functions.
The function is a wrapper for the function \code{\link[mboost]{mboost}}
in the model-based boosting package \code{mboost} via its functional pendant
\code{\link[FDboost]{FDboost}} from the \code{FDboost} package.
For manifold valued responses, like shapes, appropriate loss functions
can be fitted using an \code{\link[mfFamily]{mfFamily}} implementing the
geometric structure of the response efficiently via an
\code{\link[mfGeometry]{mfGeometry}}.
}
\details{
While the response observations \eqn{y_i} and the corresponding predictions
\eqn{\mu_i} might live on a non-linear manifold \eqn{M}, they are modeled with
an additive predictor \eqn{\eta_i} living in a linear space by
\deqn{\mu_i = g_p(\eta_i) = g_p(\sum_j h_j(x_i))}
where \eqn{g_p} is a response function, which might depend on a pole \eqn{p \in M}
and is typically chosen as \eqn{g_p = Exp_p}, the manifold exponential function
at \eqn{p}. The additive predictor \eqn{\eta_i} is composed of partial effects
\eqn{h_j(x_i)} as provided by the R packages \code{mboost} and \code{FDboost},
but potentially constraining them to a respective linear subspace, e.g.
corresponding to the tangent space at \eqn{p}.

For further details on available covariate effects see
\code{\link[FDboost]{FDboost}} and the \code{\link[baselearners]{baselearners}}
help of \code{mboost}.

Computationally, it might make a huge difference whether the response observations
are measured on a common regular grid or on irregular individual grids. In the
regular case, the linear array model can be utilized (see ...) for the design
matrix and the regular structure might also be used to speed up pole and
gradient computation.
This distinction is reflected - and controled by - the data format and
in particular the format the response is provided in. For a data set with
\eqn{N} observations, \code{data} should be
provided as data frame containing covariate columns (with \eqn{N} rows) or
as a list of these. The response should be contained as follows:
\itemize{
\item in the regular case, \code{data} has to be a list containing the response
as \code{\link[tbl_cube]{tbl_cube}} with the
\code{obj.formula} and \code{dim} variables as dimensions and the response
values as measures, listed according to the covariates.
\item in the irregular case, response can be given as either
\itemize{
\item a list of \code{data.frame}s for each observation in long format,
i.e. with one column for the response measurements and one for each variable
in \code{dim} and \code{obj.formula}. If the name of the response measurements
does not correspond to the name of the response in \code{formula}, the name
can be specified on the left-hand side of the \code{obj.formula} (or \code{dim}).
\item a list of \code{tbl_cube}s corresponding to the regular case,
but with one measure each.
\item as a single \code{data.frame} including an additional ID variable indicating
the 'row number' in the covariates. In this case, an extra \code{id} formula
has to be specified as \code{FDboost} and \code{data} has to be a list.
}
\item in the format passed to \code{FDboost}, i.e., \code{data} is a list
and the response is provided in seperate list elements:
\itemize{
\item in the regular case, the response measurements are provided as matrix
with \eqn{N} rows corresponding to the observations and the columns
containing all measurements in long format and the \code{dim} and \code{obj.formula}
variables as vectors along the columns of the matrix.
\item in the irregular case, the columns of the last irregular option above
are just separately added to the list.
}
}
}
