% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfGeometry.R
\name{mfGeometry}
\alias{mfGeometry}
\title{R6 Class implementing manifold geometries}
\description{
R6 Class implementing manifold geometries

R6 Class implementing manifold geometries
}
\details{
An abstract \code{mfGeometry} class implementing a set of methods, including
e.g. Exp- and Log-mappings, that are needed to define \code{mfboost} families
in a concise and modular way.
The arguments of the methods are vectors, matrices
or arrays possibly holding additional attributes. Besides these methods, objects
of class \code{mfGeometry} offer an active slot for the 'pole' which serves
as prototype for elements of the manifold. Children of this class
specify the geometry of interest, such as \code{\link{mfGeomEuclidean}},
\code{\link{mfGeomSphere}}, \code{\link{mfGeomPlanarShape}}, and
\code{\link{mfGeomPlanarSizeShape}}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{y_}}{slot for a data object in the internal format}

\item{\code{pole_}}{slot for a data object in the internal format}

\item{\code{weights_}}{slot for inner product weights in the internal format}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mfGeometry-structure}{\code{mfGeometry$structure()}}
\item \href{#method-mfGeometry-unstructure}{\code{mfGeometry$unstructure()}}
\item \href{#method-mfGeometry-unstructure_weights}{\code{mfGeometry$unstructure_weights()}}
\item \href{#method-mfGeometry-align}{\code{mfGeometry$align()}}
\item \href{#method-mfGeometry-register}{\code{mfGeometry$register()}}
\item \href{#method-mfGeometry-distance}{\code{mfGeometry$distance()}}
\item \href{#method-mfGeometry-exp}{\code{mfGeometry$exp()}}
\item \href{#method-mfGeometry-log}{\code{mfGeometry$log()}}
\item \href{#method-mfGeometry-transport}{\code{mfGeometry$transport()}}
\item \href{#method-mfGeometry-innerprod}{\code{mfGeometry$innerprod()}}
\item \href{#method-mfGeometry-get_normal}{\code{mfGeometry$get_normal()}}
\item \href{#method-mfGeometry-validate}{\code{mfGeometry$validate()}}
\item \href{#method-mfGeometry-clone}{\code{mfGeometry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-structure"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-structure}{}}}
\subsection{Method \code{structure()}}{
structure a vector containing the elements of the response
\code{y} in long format in the internal format of the geometry.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$structure(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{a numeric vector containing the elements of an object y in long format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-unstructure"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-unstructure}{}}}
\subsection{Method \code{unstructure()}}{
inverse operation of \code{structure}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$unstructure(y_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{an object in the internal format of the geometry}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-unstructure_weights"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-unstructure_weights}{}}}
\subsection{Method \code{unstructure_weights()}}{
convert internal inner product weights to a weight vector
matching the un-structered response \code{y} in long format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$unstructure_weights(weights_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weights_}}{inner product weights in the internal format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-align"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-align}{}}}
\subsection{Method \code{align()}}{
a function aligning \code{y_} to \code{y0_} primarly intended
for quotient geometries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$align(y_, y0_ = private$.pole_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{an object in the internal format of the geometry}

\item{\code{y0_}}{an object in the internal format of the geometry}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-register"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-register}{}}}
\subsection{Method \code{register()}}{
a function registering \code{y_} to a proper subset, say to
a subspace or an embedded submanifold.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$register(y_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{an object in the internal format of the geometry}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-distance"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-distance}{}}}
\subsection{Method \code{distance()}}{
computes the distance or a vector of distances between objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$distance(y0_, y1_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{an object in the internal format of the geometry}

\item{\code{y1_}}{an object in the internal format of the geometry}

\item{\code{...}}{other arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-exp"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-exp}{}}}
\subsection{Method \code{exp()}}{
the Riemannian Exp map mapping a tangent vector \code{v_}
at \code{y_} to the manifold.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$exp(v_, y0_ = private$.pole_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v_}}{an object repesenting a tangent vector (internal format)}

\item{\code{y0_}}{an object in the internal format of the geometry}

\item{\code{...}}{other arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-log"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-log}{}}}
\subsection{Method \code{log()}}{
the inverse of the Riemannian Exp map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$log(y_, y0_ = private$.pole_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{an object in the internal format of the geometry}

\item{\code{y0_}}{an object in the internal format of the geometry}

\item{\code{...}}{other arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-transport"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-transport}{}}}
\subsection{Method \code{transport()}}{
parallel transport of a tangent vector \code{v0_}
at \code{y0_} to the tangent space at \code{y1_}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$transport(v0_, y0_, y1_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v0_}}{an object repesenting a tangent vector (internal format)}

\item{\code{y0_}}{an object in the internal format of the geometry}

\item{\code{y1_}}{an object in the internal format of the geometry}

\item{\code{...}}{other arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-innerprod"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-innerprod}{}}}
\subsection{Method \code{innerprod()}}{
computes the inner product or a vector of inner products
between objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$innerprod(v0_, v1_ = v0_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v0_}}{an object repesenting a tangent vector (internal format)}

\item{\code{v1_}}{an object repesenting a tangent vector (internal format)}

\item{\code{...}}{other arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-get_normal"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-get_normal}{}}}
\subsection{Method \code{get_normal()}}{
returns the normal vector (in unstructured format) to the
tangent space at a point \code{y0_} considered as a linear subspace.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$get_normal(y0_ = private$.pole_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{an object in the internal format of the geometry}

\item{\code{...}}{other arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-validate"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-validate}{}}}
\subsection{Method \code{validate()}}{
function for validating structure of internal object \code{y_}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$validate(y_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{an object in the internal format of the geometry}

\item{\code{...}}{other arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfGeometry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mfGeometry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeometry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
