% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfGeometry.R
\name{mfGeomPlanarShape}
\alias{mfGeomPlanarShape}
\title{Planar Shape Space Geometry}
\description{
Geometry of the 'classic' shape space identifying planar shapes
with a centered and scaled complex vector on the complex sphere / complex
projective space.
}
\section{Super classes}{
\code{\link[manifoldboost:mfGeometry]{manifoldboost::mfGeometry}} -> \code{\link[manifoldboost:mfGeomEuclidean]{manifoldboost::mfGeomEuclidean}} -> \code{\link[manifoldboost:mfGeomUnitSphere]{manifoldboost::mfGeomUnitSphere}} -> \code{mfGeomPlanarShape}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mfGeomPlanarShape$new()}}
\item \href{#method-structure}{\code{mfGeomPlanarShape$structure()}}
\item \href{#method-unstructure}{\code{mfGeomPlanarShape$unstructure()}}
\item \href{#method-unstructure_weights}{\code{mfGeomPlanarShape$unstructure_weights()}}
\item \href{#method-structure_weights}{\code{mfGeomPlanarShape$structure_weights()}}
\item \href{#method-register}{\code{mfGeomPlanarShape$register()}}
\item \href{#method-register_v}{\code{mfGeomPlanarShape$register_v()}}
\item \href{#method-log}{\code{mfGeomPlanarShape$log()}}
\item \href{#method-transport}{\code{mfGeomPlanarShape$transport()}}
\item \href{#method-plot}{\code{mfGeomPlanarShape$plot()}}
\item \href{#method-validate}{\code{mfGeomPlanarShape$validate()}}
\item \href{#method-get_normal}{\code{mfGeomPlanarShape$get_normal()}}
\item \href{#method-clone}{\code{mfGeomPlanarShape$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeometry" data-id="distance">}\href{../../manifoldboost/html/mfGeometry.html#method-distance}{\code{manifoldboost::mfGeometry$distance()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="align">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-align}{\code{manifoldboost::mfGeomEuclidean$align()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="innerprod">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-innerprod}{\code{manifoldboost::mfGeomEuclidean$innerprod()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomUnitSphere" data-id="exp">}\href{../../manifoldboost/html/mfGeomUnitSphere.html#method-exp}{\code{manifoldboost::mfGeomUnitSphere$exp()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize planar shape geometry for data given
in (long) FDboost format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$new(data, formula, weight_fun = NULL, arg_range = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data in the format used in \code{\link{FDboost}}.}

\item{\code{formula}}{formula describing the internal structure of the data,
intended for the \code{obj.formula} of \code{mfboost}
(interpreted by \code{mfInterpret_objformula}).}

\item{\code{weight_fun}}{a function computing inner product weights for the geometry
taking two arguments: \code{arg}, the argument of the functional data
specified in \code{formula}, \code{range} the range of \code{arg}.}

\item{\code{arg_range}}{the \code{range} supplied to the \code{weight_fun}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-structure"></a>}}
\if{latex}{\out{\hypertarget{method-structure}{}}}
\subsection{Method \code{structure()}}{
convert 2D vectors (given in long format with dimension
indicator) to complex
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$structure(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{a numeric vector representing an element of the manifold in an
unstructured way. If \code{y} is of length \eqn{l}, the first \eqn{l/2} elements
of \code{y} correspond to the real part and the remaining \eqn{l/2} elements to
imaginary part of the complex vector \code{y_}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unstructure"></a>}}
\if{latex}{\out{\hypertarget{method-unstructure}{}}}
\subsection{Method \code{unstructure()}}{
convert complex back to long format numeric
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$unstructure(y_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unstructure_weights"></a>}}
\if{latex}{\out{\hypertarget{method-unstructure_weights}{}}}
\subsection{Method \code{unstructure_weights()}}{
double inner product weights for complex vectors
to appear in both dimensions in numeric long format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$unstructure_weights(weights_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-structure_weights"></a>}}
\if{latex}{\out{\hypertarget{method-structure_weights}{}}}
\subsection{Method \code{structure_weights()}}{
remove double weights to obtain
inner product weights for complex vectors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$structure_weights(weights)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register"></a>}}
\if{latex}{\out{\hypertarget{method-register}{}}}
\subsection{Method \code{register()}}{
center \code{y_} and scale it to unit norm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$register(y_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register_v"></a>}}
\if{latex}{\out{\hypertarget{method-register_v}{}}}
\subsection{Method \code{register_v()}}{
orthogonally project \code{v_} into the tangent space of \code{y0_}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$register_v(v_, y0_ = self$pole_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v_}}{a complex tangent vector.}

\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
Apply Log function of the sphere after rotation alignment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$log(
  y_,
  y0_ = self$pole_,
  method = c("simple", "alternative")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a complex vector on the sphere, orthogonal to the constant.}

\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transport"></a>}}
\if{latex}{\out{\hypertarget{method-transport}{}}}
\subsection{Method \code{transport()}}{
for the parallel transport previous alignment is assumed,
such that the transport of the sphere is directly applied.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$transport(
  v0_,
  y0_,
  y1_,
  method = c("horizontal", "simple", "general")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v0_}}{a complex tangent vector.}

\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
default plotting function for planar shapes, plotting \code{y_}
in front of \code{y0_} (after alignment).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$plot(
  y_ = self$y_,
  y0_ = self$pole_,
  ylab = NA,
  xlab = NA,
  col = "black",
  ylim = range(Im(c(y_, y0_))),
  xlim = range(Re(c(y_, y0_))),
  pch = 19,
  asp = 1,
  yaxt = "n",
  xaxt = "n",
  type = "p",
  y0_par = list(col = "darkgrey", type = type),
  seg_par = list(col = "grey"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a complex vector on the sphere, orthogonal to the constant.}

\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}

\item{\code{ylab, xlab, xlim, ylim, xaxt, yaxt}}{graphical parameters passed to \code{base::plot}
with modified defaults.}

\item{\code{col, pch, type}}{graphical parameters passed to \code{base::plot} referring to \code{y_}.}

\item{\code{seg_par}}{graphical parameters for line segments connecting \code{y_}
and \code{y0_}.}

\item{\code{...}}{other arguments passed to \code{base::plot}.}

\item{\code{y0_param}}{graphical parameters for \code{y0_}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$validate(y0_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_normal"></a>}}
\if{latex}{\out{\hypertarget{method-get_normal}{}}}
\subsection{Method \code{get_normal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$get_normal(y0_ = self$pole_, weighted = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{a complex vector on the sphere, orthogonal to the constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomPlanarShape$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
