% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfGeometry.R
\name{mfGeomProduct}
\alias{mfGeomProduct}
\title{A general product geometry}
\description{
An \code{mfGeometry} class implementing a product geometry in an abstract
general way. While there might be more efficient implementations using the
particular structure of your geometry of interest, this should work for
products of any mfGeometries.
}
\section{Super class}{
\code{\link[manifoldboost:mfGeometry]{manifoldboost::mfGeometry}} -> \code{mfGeomProduct}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mfGeomProduct$new()}}
\item \href{#method-slice}{\code{mfGeomProduct$slice()}}
\item \href{#method-structure}{\code{mfGeomProduct$structure()}}
\item \href{#method-unstructure}{\code{mfGeomProduct$unstructure()}}
\item \href{#method-unstructure_weights}{\code{mfGeomProduct$unstructure_weights()}}
\item \href{#method-structure_weights}{\code{mfGeomProduct$structure_weights()}}
\item \href{#method-align}{\code{mfGeomProduct$align()}}
\item \href{#method-register}{\code{mfGeomProduct$register()}}
\item \href{#method-register_v}{\code{mfGeomProduct$register_v()}}
\item \href{#method-distance}{\code{mfGeomProduct$distance()}}
\item \href{#method-exp}{\code{mfGeomProduct$exp()}}
\item \href{#method-log}{\code{mfGeomProduct$log()}}
\item \href{#method-transport}{\code{mfGeomProduct$transport()}}
\item \href{#method-innerprod}{\code{mfGeomProduct$innerprod()}}
\item \href{#method-plot}{\code{mfGeomProduct$plot()}}
\item \href{#method-get_normal}{\code{mfGeomProduct$get_normal()}}
\item \href{#method-validate}{\code{mfGeomProduct$validate()}}
\item \href{#method-clone}{\code{mfGeomProduct$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$new(mfGeom_default, data, formula)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-slice"></a>}}
\if{latex}{\out{\hypertarget{method-slice}{}}}
\subsection{Method \code{slice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$slice(which)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-structure"></a>}}
\if{latex}{\out{\hypertarget{method-structure}{}}}
\subsection{Method \code{structure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$structure(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{a list of \code{y} elements of the component manifolds in an
unstructured 'flattened' way}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unstructure"></a>}}
\if{latex}{\out{\hypertarget{method-unstructure}{}}}
\subsection{Method \code{unstructure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$unstructure(y_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a list of objects living in the comp. manifolds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unstructure_weights"></a>}}
\if{latex}{\out{\hypertarget{method-unstructure_weights}{}}}
\subsection{Method \code{unstructure_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$unstructure_weights(weights_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-structure_weights"></a>}}
\if{latex}{\out{\hypertarget{method-structure_weights}{}}}
\subsection{Method \code{structure_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$structure_weights(weights)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-align"></a>}}
\if{latex}{\out{\hypertarget{method-align}{}}}
\subsection{Method \code{align()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$align(y_, y0_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a list of objects living in the comp. manifolds}

\item{\code{y0_}}{a list of objects living in the comp. manifolds}

\item{\code{...}}{other arguments passed to underlying \code{mfGeometry}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register"></a>}}
\if{latex}{\out{\hypertarget{method-register}{}}}
\subsection{Method \code{register()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$register(y_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a list of objects living in the comp. manifolds}

\item{\code{...}}{other arguments passed to underlying \code{mfGeometry}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register_v"></a>}}
\if{latex}{\out{\hypertarget{method-register_v}{}}}
\subsection{Method \code{register_v()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$register_v(v_, y0_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v_}}{a list of tangent vectors living in the comp. manifolds}

\item{\code{y0_}}{a list of objects living in the comp. manifolds}

\item{\code{...}}{other arguments passed to underlying \code{mfGeometry}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distance"></a>}}
\if{latex}{\out{\hypertarget{method-distance}{}}}
\subsection{Method \code{distance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$distance(y0_, y1_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{a list of objects living in the comp. manifolds}

\item{\code{y1_}}{a list of objects living in the comp. manifolds}

\item{\code{...}}{other arguments passed to underlying \code{mfGeometry}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exp"></a>}}
\if{latex}{\out{\hypertarget{method-exp}{}}}
\subsection{Method \code{exp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$exp(v_, y0_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v_}}{a list of tangent vectors living in the comp. manifolds}

\item{\code{y0_}}{a list of objects living in the comp. manifolds}

\item{\code{...}}{other arguments passed to underlying \code{mfGeometry}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$log(y_, y0_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a list of objects living in the comp. manifolds}

\item{\code{y0_}}{a list of objects living in the comp. manifolds}

\item{\code{...}}{other arguments passed to underlying \code{mfGeometry}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transport"></a>}}
\if{latex}{\out{\hypertarget{method-transport}{}}}
\subsection{Method \code{transport()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$transport(v0_, y0_, y1_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v0_}}{a list of tangent vectors living in the comp. manifolds}

\item{\code{y0_}}{a list of objects living in the comp. manifolds}

\item{\code{y1_}}{a list of objects living in the comp. manifolds}

\item{\code{...}}{other arguments passed to underlying \code{mfGeometry}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-innerprod"></a>}}
\if{latex}{\out{\hypertarget{method-innerprod}{}}}
\subsection{Method \code{innerprod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$innerprod(v0_, v1_ = v0_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v0_}}{a list of tangent vectors living in the comp. manifolds}

\item{\code{v1_}}{a list of tangent vectors living in the comp. manifolds}

\item{\code{...}}{other arguments passed to underlying \code{mfGeometry}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$plot(y_ = self$y_, y0_ = self$pole_, main = names(y_), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a list of objects living in the comp. manifolds}

\item{\code{y0_}}{a list of objects living in the comp. manifolds}

\item{\code{...}}{other arguments passed to underlying \code{mfGeometry}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_normal"></a>}}
\if{latex}{\out{\hypertarget{method-get_normal}{}}}
\subsection{Method \code{get_normal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$get_normal(y0_ = self$pole_, weighted = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{a list of objects living in the comp. manifolds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$validate(y0_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{a list of objects living in the comp. manifolds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomProduct$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
