% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfGeometry.R
\name{mfGeomUnitSphere}
\alias{mfGeomUnitSphere}
\title{Geometry of the Unit Sphere}
\description{
The geometry of the n-dimensional unit sphere, treating the complex sphere is,
in case, as a real manifold.
}
\section{Super classes}{
\code{\link[manifoldboost:mfGeometry]{manifoldboost::mfGeometry}} -> \code{\link[manifoldboost:mfGeomEuclidean]{manifoldboost::mfGeomEuclidean}} -> \code{mfGeomUnitSphere}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-register}{\code{mfGeomUnitSphere$register()}}
\item \href{#method-register_v}{\code{mfGeomUnitSphere$register_v()}}
\item \href{#method-log}{\code{mfGeomUnitSphere$log()}}
\item \href{#method-exp}{\code{mfGeomUnitSphere$exp()}}
\item \href{#method-transport}{\code{mfGeomUnitSphere$transport()}}
\item \href{#method-get_normal}{\code{mfGeomUnitSphere$get_normal()}}
\item \href{#method-clone}{\code{mfGeomUnitSphere$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeometry" data-id="distance">}\href{../../manifoldboost/html/mfGeometry.html#method-distance}{\code{manifoldboost::mfGeometry$distance()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeometry" data-id="unstructure_weights">}\href{../../manifoldboost/html/mfGeometry.html#method-unstructure_weights}{\code{manifoldboost::mfGeometry$unstructure_weights()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="align">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-align}{\code{manifoldboost::mfGeomEuclidean$align()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="innerprod">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-innerprod}{\code{manifoldboost::mfGeomEuclidean$innerprod()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="structure">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-structure}{\code{manifoldboost::mfGeomEuclidean$structure()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="unstructure">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-unstructure}{\code{manifoldboost::mfGeomEuclidean$unstructure()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="manifoldboost" data-topic="mfGeomEuclidean" data-id="validate">}\href{../../manifoldboost/html/mfGeomEuclidean.html#method-validate}{\code{manifoldboost::mfGeomEuclidean$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register"></a>}}
\if{latex}{\out{\hypertarget{method-register}{}}}
\subsection{Method \code{register()}}{
\code{y} vectors are scaled to unit norm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomUnitSphere$register(y_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a numeric/complex vector on the sphere}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register_v"></a>}}
\if{latex}{\out{\hypertarget{method-register_v}{}}}
\subsection{Method \code{register_v()}}{
vectors \code{v}
are orthogonalized with respect to \code{y0_} guaranteeing proper tangent
vectors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomUnitSphere$register_v(v_, y0_ = private$.pole_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v_}}{a numeric/complex tangent vector}

\item{\code{y0_}}{a numeric/complex vector on the sphere}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomUnitSphere$log(
  y_,
  y0_ = private$.pole_,
  method = c("simple", "alternative")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_}}{a numeric/complex vector on the sphere}

\item{\code{y0_}}{a numeric/complex vector on the sphere}

\item{\code{method}}{character string, for choosing one of two slightly different
methods to implement log-method ("simple" or "alternative").}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exp"></a>}}
\if{latex}{\out{\hypertarget{method-exp}{}}}
\subsection{Method \code{exp()}}{
moving \code{v_} along an arc on the sphere.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomUnitSphere$exp(v_, y0_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v_}}{a numeric/complex tangent vector}

\item{\code{y0_}}{a numeric/complex vector on the sphere}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transport"></a>}}
\if{latex}{\out{\hypertarget{method-transport}{}}}
\subsection{Method \code{transport()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomUnitSphere$transport(
  v0_,
  y0_,
  y1_,
  method = c("general", "horizontal", "simple")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v0_}}{a numeric/complex tangent vector}

\item{\code{y0_}}{a numeric/complex vector on the sphere}

\item{\code{y1_}}{a numeric/complex vector on the sphere}

\item{\code{method}}{expression used for parallel transport: "general" corresponds to the one of
Cornea et al. 2017, "horizontal" to the one of Dryden & Mardia 2012 p. 76,
and "simple" to a simplified version of "general" where \code{v0_}
is horizontal to \code{y0_}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_normal"></a>}}
\if{latex}{\out{\hypertarget{method-get_normal}{}}}
\subsection{Method \code{get_normal()}}{
normal vector just corresponds to the pole itself
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomUnitSphere$get_normal(y0_ = private$.pole_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y0_}}{a numeric/complex vector on the sphere}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfGeomUnitSphere$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
