% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfFamily.R
\name{mfPole}
\alias{mfPole}
\title{R6 Class for mfboost pole model object for use in \code{mfboost_family}}
\description{
R6 Class for mfboost pole model object for use in \code{mfboost_family}

R6 Class for mfboost pole model object for use in \code{mfboost_family}
}
\details{
A pole (similar to an offset) for a \code{mfboost} model is typically result
of some kind of model fit, which might be another simpler mfboost model
or some other kind of model procedure estimating the pole from the data.
For later model predictions on new data, it is important to store the
not only the resulting pole, but the whole model behind it, to have access
to new pole predictions. The abstract R6 class \code{mfPole} implements
the basic structure needed to provide a pole for mfboost.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{A fitted model object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mfPole-fit}{\code{mfPole$fit()}}
\item \href{#method-mfPole-predict}{\code{mfPole$predict()}}
\item \href{#method-mfPole-clone}{\code{mfPole$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfPole-fit"></a>}}
\if{latex}{\out{\hypertarget{method-mfPole-fit}{}}}
\subsection{Method \code{fit()}}{
An initialization function determining the pole from the
data. The function takes two arguments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfPole$fit(obj.formula, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj.formula}}{Model formula for the pole model.}

\item{\code{data}}{Data in \code{FDboost} format.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfPole-predict"></a>}}
\if{latex}{\out{\hypertarget{method-mfPole-predict}{}}}
\subsection{Method \code{predict()}}{
A function taking \code{newdata} in the same format as \code{data}
and generating a prediction as a simple vector fitting to the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfPole$predict(newdata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{in \code{FDboost} format. The default \code{NULL}
specifies the original data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mfPole-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mfPole-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfPole$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
