% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfPoles.R
\name{planarshape_full_proc}
\alias{planarshape_full_proc}
\title{Full Procrustres mean of planar shapes}
\usage{
planarshape_full_proc(
  formula,
  data,
  cyclic = FALSE,
  smoothed.cov = FALSE,
  cov.k = 10,
  arg.grid.len = 50,
  weights = NULL,
  arg.range = NULL,
  mf = NULL,
  mfboost.return = FALSE
)
}
\arguments{
\item{formula}{a formula in the form \code{shape ~ value^dim ~ blearner(arg) | id},
where \code{blearner} is typically \code{bbs} if \code{arg} is continuous and \code{bols} if discrete.}

\item{data}{data containing the shape in a format convertible to \code{shape_dat}.
Shape should be contained as \code{tbl_cube} (\code{shape_cube}) if data is regular.}

\item{cyclic}{logical, should cyclic B-splines be used?}

\item{smoothed.cov}{logical, should covariance matrix be smoothed?}

\item{cov.k}{integer, number of basis elements for each margin in covariance smoothing.}

\item{arg.grid.len}{integer, grid length for evaluation of smoothed covariance.}

\item{weights}{numeric weight vector.}

\item{mf}{\code{mfGeometry} object.}

\item{mfboost.return}{logocal, should results be returned in the format needed
for mfboost?}
}
\value{
in the default setting, the estimated full procrustres mean in
shape_frame_long format
(data.frame with coordinates as real values stored in a single column).
}
\description{
EXPERIMENTAL. CURRENTLY NOT EXPORTED!
}
