% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfboost.R
\name{predict.mfboost}
\alias{predict.mfboost}
\title{Predict with mfboost model}
\usage{
\method{predict}{mfboost}(
  object,
  newdata = NULL,
  which = NULL,
  type = c("response", "link"),
  raw = FALSE,
  ...
)
}
\arguments{
\item{object}{object of class \code{mfboost} for plotting}

\item{newdata}{an optional dataset with new data for prediction. Should be
of the same structure as the original dataset
(also including a slot for the response!). The default NULL returns predictions
on the data used for fitting.}

\item{which}{a subset of base-learners to take into account for computing
predictions or coefficients as in \code{predict.mboost}. \code{which = 0} will
predict the pole.}

\item{...}{other arguments passed to \code{predict.mboost}.}
}
\value{

}
\description{
Get predictions of an \code{mfboost} model.
}
